Refactor to have Modem send frames to decoder
