+++ .gitignore
@@ -1,6 +1,7 @@
 cmd/m17-text-cli/m17-text-cli
 cmd/m17-gateway/m17-gateway
 cmd/m17-gateway/m17-gateway.ini
+cmd/m17-gateway/M17Hosts.txt
 cmd/m17-gateway/build/
 cmd/m17-message/m17-message
 cmd/m17-message/m17-message.app/
